upload-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
boto3
kafka-python
pydantic
python-multipart
